move out AppNavHost composable to a different file from MainActivity
